Refactor game logic from Board.jsx to OnlineGameView
