+++ App.js
@@ -2,7 +2,7 @@ import React from "react";
 
 import { StatusBar } from "expo-status-bar";
 import { StyleSheet } from "react-native";
-import { Board } from "./components/Board/Board.jsx";
+import Board from "./components/Board/Board.jsx";
 import Main from "./components/Main/Main.jsx";
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
@@ -11,7 +11,8 @@ import Login from "./components/Login/Login.jsx";
 import Tutorial from "./components/Tutorial/Tutorial.jsx";
 import Stats from "./components/Profile/Stats.jsx";
 import GameStat from "./components/Profile/GameStat.jsx";
-import GameReview from "./views/GameReview.jsx";
+import OnlineGameView from "./views/gameView/onlineGameView/OnlineGameView.jsx";
+//import GameReview from "./views/GameReview.jsx";
 const Stack = createStackNavigator();
 
 export default function App() {
@@ -24,13 +25,13 @@ export default function App() {
           component={Board}
           initialParams={{ isLocalGame: true }}
         />
-        <Stack.Screen name="onlineGame" component={Board} />
+        <Stack.Screen name="onlineGame" component={OnlineGameView} />
         <Stack.Screen name="login" component={Login} />
         <Stack.Screen name="register" component={Register} />
         <Stack.Screen name="tutorial" component={Tutorial} />
         <Stack.Screen name="profile" component={Stats} />
         <Stack.Screen name="stats" component={GameStat} />
-        <Stack.Screen name="reviewGame" component={GameReview} />
+        {/* <Stack.Screen name="reviewGame" component={GameReview} /> */}
       </Stack.Navigator>
       <StatusBar style="auto" />
     </NavigationContainer>
