+++ views/gameView/gameReview/GameReview.jsx
@@ -1,0 +1,1 @@
+export default {};
